requirements.txt
flask==3.0.0
GitPython==3.1.40
PyYAML==6.0.1
jsonschema==4.22.0
requests==2.31.0
python-dateutil==2.8.2